Use new `mint::When` enum. instead of `bool force`
`bool force` was equivalent to choosing between `mint::When::Auto`
and `mint::When::Always`.

`mint::When` also offers `mint::When::Never` which forces mint::mint()
to _not_ emit SGR codes.

This is needed in case, for some reason, hasTerminalSupport() believes
that the connected terminal supports attributes even if it's not
the case.

One can also wrap mint::mint() to set the `when` parameter from an
environment variable.

I'm also adding the current version to the header (since it's meant to
be a volatile header) and sorting the includes.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>