Add mint::escapeAnsi(): remove SGR codes from a string
This new function isn't strictly related to mint, but it can prove
useful in any mint app context: it removes the SGR codes from a string
which mint::mint() can produce.

This is useful if you already have a mint-formatted string, for example
to know its true content width on the terminal:

    mint::escapeAnsi(myMintStr).size()

Also adding tests and updating the documentation.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>